Add files via upload
Uses HSV thresholding rather than grayscale so that it can ignore the leaves of the tree and the manipulator when thresholding, rather than when doing polygon estimation. HSV thresholding allows for easier thresholding against different backgrounds, and makes it easier to single out one color. In this case, we are assuming all apples are bright red. Uses contour detection and polygon estimation to aproximate how many sides each apple has. If it has less than 5 sides, it is considered a bad apple, and if it has more than 5, it is a good apple.

Slightly changed Robot 3.0 so that tree is immediately in view of the camera and so that the shapes for good and bad apples are a sphere for good, and a triangular pyramid and a cone for bad apples. Slightly adjusted the camera pole so that it wasn't in the camera's view.